creates Database folder for Firebase interactions
