+++ src/App.test.js
@@ -4,7 +4,7 @@ import App from './App';
 import userEvent from '@testing-library/user-event';
 import { MemoryRouter } from 'react-router-dom';
 import { AuthProvider } from './Auth';
-import { auth, db } from './Firebase';
+import { auth, db } from './Database/Firebase';
 
 jest.mock('./Firebase');
