+++ src/RecordIncidentForm.js
@@ -3,7 +3,7 @@ import Button from 'react-bootstrap/Button';
 import Col from 'react-bootstrap/Col';
 import Form from 'react-bootstrap/Form';
 import Row from 'react-bootstrap/Row';
-import { addIncident, getParticipantById } from './FirebaseIncidentForm';
+import { addIncident, getParticipantById } from './Database';
 
 export default () => {
   const [state, update] = useReducer(
